+++ packages/atomic-react/CHANGELOG.md
@@ -1,3 +1,9 @@
+## <small>3.3.11 (2025-09-24)</small>
+
+* chore(query summary): migrate atomic-query-summary (#6025) ([25218e2](https://github.com/coveo/ui-kit/commits/25218e2)), closes [#6025](https://github.com/coveo/ui-kit/issues/6025)
+
+
+
 ## <small>3.3.10 (2025-09-18)</small>
 
 * [Version Bump][skip ci]: ui-kit publish ([a3fb13e](https://github.com/coveo/ui-kit/commits/a3fb13e))